Configuration Doesn't save TimeZones that are not integer!
git-svn-id: http://joomlacode.org/svn/joomla/development/trunk@8341 6f6e1ebd-4c2b-0410-823f-f34bde69bce9

+++ administrator/components/com_config/controllers/application.php
@@ -277,7 +277,7 @@ class ConfigControllerApplication extends ConfigController
 		$config_array['tmp_path']			= JRequest::getVar('tmp_path', JPATH_ROOT.DS.'tmp', 'post', 'string');
 
 		// LOCALE SETTINGS
-		$config_array['offset']				= JRequest::getVar('offset', 0, 'post', 'int');
+		$config_array['offset']				= JRequest::getVar('offset', 0, 'post', 'float');
 
 		// CACHE SETTINGS
 		$config_array['caching']			= JRequest::getVar('caching', 0, 'post', 'int');
